refactor: extract modules, deduplicate handlers, harden page summary
- Extract pierce-shadow functions to src/background/pierce-shadow.ts
- Extract DOM dispatcher utilities to src/background/dom-dispatch.ts
- Extract buildBrowserActionArgs helper in tools_interact.go (dedup 4 blocks)
- Remove string "true"/"false" normalization from parsePierceShadowInput
- Strip control characters from cleanText in page summary script
- Add mode validation guard to pageSummaryScript IIFE

File sizes: dom-primitives.ts 1162→893, pending-queries.ts 1371→1308,
tools_interact.go 768→736
