+++ extension/background/dom-dispatch.d.ts
@@ -1,0 +1,7 @@
+import type { PendingQuery } from '../types/queries';
+import type { SyncClient } from './sync-client';
+type SendAsyncResult = (syncClient: SyncClient, queryId: string, correlationId: string, status: 'complete' | 'error' | 'timeout', result?: unknown, error?: string) => void;
+type ActionToast = (tabId: number, text: string, detail?: string, state?: 'trying' | 'success' | 'warning' | 'error', durationMs?: number) => void;
+export declare function executeDOMAction(query: PendingQuery, tabId: number, syncClient: SyncClient, sendAsyncResult: SendAsyncResult, actionToast: ActionToast): Promise<void>;
+export {};
+//# sourceMappingURL=dom-dispatch.d.ts.map
