+++ extension/background/dom-primitives.d.ts
@@ -1,5 +1,3 @@
-import type { PendingQuery } from '../types/queries';
-import type { SyncClient } from './sync-client';
 interface DOMResult {
     success: boolean;
     action: string;
@@ -19,8 +17,6 @@ interface DOMResult {
     };
     analysis?: string;
 }
-type SendAsyncResult = (syncClient: SyncClient, queryId: string, correlationId: string, status: 'complete' | 'error' | 'timeout', result?: unknown, error?: string) => void;
-type ActionToast = (tabId: number, text: string, detail?: string, state?: 'trying' | 'success' | 'warning' | 'error', durationMs?: number) => void;
 /**
  * Single self-contained function for all DOM primitives.
  * Passed to chrome.scripting.executeScript({ func: domPrimitive, args: [...] }).
@@ -50,6 +46,5 @@ export declare function domWaitFor(selector: string, timeoutMs: number): Promise
 export declare function domFrameProbe(frameTarget: string | number): {
     matches: boolean;
 };
-export declare function executeDOMAction(query: PendingQuery, tabId: number, syncClient: SyncClient, sendAsyncResult: SendAsyncResult, actionToast: ActionToast): Promise<void>;
 export {};
 //# sourceMappingURL=dom-primitives.d.ts.map
