+++ src/background/dom-dispatch.ts
@@ -1,0 +1,286 @@
+// dom-dispatch.ts — DOM action dispatcher and utilities.
+// Extracted from dom-primitives.ts to reduce file size.
+// Script builders (domPrimitive, domWaitFor, domFrameProbe) stay in dom-primitives.ts
+// because they must be self-contained for chrome.scripting.executeScript.
+
+import type { PendingQuery } from '../types/queries'
+import type { SyncClient } from './sync-client'
+import { domPrimitive, domWaitFor, domFrameProbe } from './dom-primitives'
+
+// Result shape returned by domPrimitive (compile-time only — erased at runtime)
+interface DOMResult {
+  success: boolean
+  action: string
+  selector: string
+  value?: unknown
+  error?: string
+  message?: string
+  dom_summary?: string
+  timing?: { total_ms: number }
+  dom_changes?: { added: number; removed: number; modified: number; summary: string }
+  analysis?: string
+}
+
+type SendAsyncResult = (
+  syncClient: SyncClient,
+  queryId: string,
+  correlationId: string,
+  status: 'complete' | 'error' | 'timeout',
+  result?: unknown,
+  error?: string
+) => void
+
+type ActionToast = (
+  tabId: number,
+  text: string,
+  detail?: string,
+  state?: 'trying' | 'success' | 'warning' | 'error',
+  durationMs?: number
+) => void
+
+interface DOMActionParams {
+  action?: string
+  selector?: string
+  text?: string
+  value?: string
+  clear?: boolean
+  checked?: boolean
+  name?: string
+  timeout_ms?: number
+  reason?: string
+  analyze?: boolean
+  frame?: string | number
+}
+
+function parseDOMParams(query: PendingQuery): DOMActionParams | null {
+  try {
+    return typeof query.params === 'string' ? JSON.parse(query.params) : (query.params as DOMActionParams)
+  } catch {
+    return null
+  }
+}
+
+function isReadOnlyAction(action: string): boolean {
+  return action === 'list_interactive' || action.startsWith('get_')
+}
+
+type DOMExecutionTarget = { tabId: number; allFrames: true } | { tabId: number; frameIds: number[] }
+
+function normalizeFrameTarget(frame: unknown): string | number | undefined | null {
+  if (frame === undefined || frame === null) return undefined
+  if (typeof frame === 'number') {
+    if (!Number.isInteger(frame) || frame < 0) return null
+    return frame
+  }
+  if (typeof frame === 'string') {
+    const trimmed = frame.trim()
+    if (trimmed.length === 0) return null
+    return trimmed
+  }
+  return null
+}
+
+async function resolveExecutionTarget(tabId: number, frame: unknown): Promise<DOMExecutionTarget> {
+  const normalized = normalizeFrameTarget(frame)
+  if (normalized === null) {
+    throw new Error('invalid_frame')
+  }
+
+  if (normalized === undefined || normalized === 'all') {
+    return { tabId, allFrames: true }
+  }
+
+  const probeResults = await chrome.scripting.executeScript({
+    target: { tabId, allFrames: true },
+    world: 'MAIN',
+    func: domFrameProbe,
+    args: [normalized]
+  })
+
+  const frameIds = Array.from(
+    new Set(
+      probeResults
+        .filter((r) => !!(r.result as { matches?: boolean } | undefined)?.matches)
+        .map((r) => r.frameId)
+        .filter((id): id is number => typeof id === 'number')
+    )
+  )
+
+  if (frameIds.length === 0) {
+    throw new Error('frame_not_found')
+  }
+
+  return { tabId, frameIds }
+}
+
+/** Pick the best result from multi-frame executeScript. Prefers main frame, falls back to first success. */
+function pickFrameResult(results: chrome.scripting.InjectionResult[]): { result: unknown; frameId: number } | null {
+  const mainFrame = results.find((r) => r.frameId === 0)
+  if (mainFrame?.result && (mainFrame.result as DOMResult).success) {
+    return { result: mainFrame.result, frameId: 0 }
+  }
+  for (const r of results) {
+    if (r.result && (r.result as DOMResult).success) {
+      return { result: r.result, frameId: r.frameId }
+    }
+  }
+  if (mainFrame?.result) return { result: mainFrame.result, frameId: 0 }
+  return results[0] ? { result: results[0].result, frameId: results[0].frameId } : null
+}
+
+/** Merge list_interactive results from all frames (up to 100 elements). */
+function mergeListInteractive(results: chrome.scripting.InjectionResult[]): { success: boolean; elements: unknown[] } {
+  const elements: unknown[] = []
+  for (const r of results) {
+    const res = r.result as { success?: boolean; elements?: unknown[] } | null
+    if (res?.elements) elements.push(...res.elements)
+    if (elements.length >= 100) break
+  }
+  return { success: true, elements: elements.slice(0, 100) }
+}
+
+async function executeWaitFor(
+  target: DOMExecutionTarget,
+  params: DOMActionParams
+): Promise<chrome.scripting.InjectionResult[] | DOMResult> {
+  const selector = params.selector || ''
+  const quickCheck = await chrome.scripting.executeScript({
+    target,
+    world: 'MAIN',
+    func: domPrimitive,
+    args: [params.action!, selector, { timeout_ms: params.timeout_ms }]
+  })
+  const quickPicked = pickFrameResult(quickCheck)
+  const quickResult = quickPicked?.result as DOMResult | undefined
+  if (quickResult?.success) return quickResult
+
+  return chrome.scripting.executeScript({
+    target,
+    world: 'MAIN',
+    func: domWaitFor,
+    args: [selector, params.timeout_ms || 5000]
+  })
+}
+
+async function executeStandardAction(
+  target: DOMExecutionTarget,
+  params: DOMActionParams
+): Promise<chrome.scripting.InjectionResult[]> {
+  return chrome.scripting.executeScript({
+    target,
+    world: 'MAIN',
+    func: domPrimitive,
+    args: [
+      params.action!,
+      params.selector || '',
+      {
+        text: params.text,
+        value: params.value,
+        clear: params.clear,
+        checked: params.checked,
+        name: params.name,
+        timeout_ms: params.timeout_ms,
+        analyze: params.analyze
+      }
+    ]
+  })
+}
+
+function sendToastForResult(
+  tabId: number,
+  readOnly: boolean,
+  result: { success?: boolean; error?: string },
+  actionToast: ActionToast,
+  toastLabel: string,
+  toastDetail: string | undefined
+): void {
+  if (readOnly) return
+  if (result.success) {
+    actionToast(tabId, toastLabel, toastDetail, 'success')
+  } else {
+    actionToast(tabId, toastLabel, result.error || 'failed', 'error')
+  }
+}
+
+// #lizard forgives
+export async function executeDOMAction(
+  query: PendingQuery,
+  tabId: number,
+  syncClient: SyncClient,
+  sendAsyncResult: SendAsyncResult,
+  actionToast: ActionToast
+): Promise<void> {
+  const params = parseDOMParams(query)
+  if (!params) {
+    sendAsyncResult(syncClient, query.id, query.correlation_id!, 'error', null, 'invalid_params')
+    return
+  }
+
+  const { action, selector, reason } = params
+  if (!action) {
+    sendAsyncResult(syncClient, query.id, query.correlation_id!, 'error', null, 'missing_action')
+    return
+  }
+  if (action === 'wait_for' && !selector) {
+    sendAsyncResult(syncClient, query.id, query.correlation_id!, 'error', null, 'missing_selector')
+    return
+  }
+
+  const toastLabel = reason || action
+  const toastDetail = reason ? undefined : selector || 'page'
+  const readOnly = isReadOnlyAction(action)
+
+  try {
+    const executionTarget = await resolveExecutionTarget(tabId, params.frame)
+    const tryingShownAt = Date.now()
+    if (!readOnly) actionToast(tabId, toastLabel, toastDetail, 'trying', 10000)
+
+    const rawResult =
+      action === 'wait_for'
+        ? await executeWaitFor(executionTarget, params)
+        : await executeStandardAction(executionTarget, params)
+
+    // wait_for quick-check can return a DOMResult directly
+    if (!Array.isArray(rawResult)) {
+      if (!readOnly) actionToast(tabId, toastLabel, toastDetail, 'success')
+      sendAsyncResult(syncClient, query.id, query.correlation_id!, 'complete', rawResult)
+      return
+    }
+
+    // Ensure "trying" toast is visible for at least 500ms
+    const MIN_TOAST_MS = 500
+    const elapsed = Date.now() - tryingShownAt
+    if (!readOnly && elapsed < MIN_TOAST_MS) await new Promise((r) => setTimeout(r, MIN_TOAST_MS - elapsed))
+
+    // list_interactive: merge elements from all frames
+    if (action === 'list_interactive') {
+      const merged = mergeListInteractive(rawResult)
+      sendAsyncResult(syncClient, query.id, query.correlation_id!, 'complete', merged)
+      return
+    }
+
+    const picked = pickFrameResult(rawResult)
+    const firstResult = picked?.result
+    if (firstResult && typeof firstResult === 'object') {
+      const resultPayload =
+        params.frame !== undefined && params.frame !== null && picked
+          ? { ...(firstResult as Record<string, unknown>), frame_id: picked.frameId }
+          : firstResult
+      sendToastForResult(
+        tabId,
+        readOnly,
+        resultPayload as { success?: boolean; error?: string },
+        actionToast,
+        toastLabel,
+        toastDetail
+      )
+      sendAsyncResult(syncClient, query.id, query.correlation_id!, 'complete', resultPayload)
+    } else {
+      if (!readOnly) actionToast(tabId, toastLabel, 'no result', 'error')
+      sendAsyncResult(syncClient, query.id, query.correlation_id!, 'error', null, 'no_result')
+    }
+  } catch (err) {
+    actionToast(tabId, action, (err as Error).message, 'error')
+    sendAsyncResult(syncClient, query.id, query.correlation_id!, 'error', null, (err as Error).message)
+  }
+}
